feat: support specifying a subset of fields to read
Being able to specify a subset of fields to read and only read
these is a particularly useful feature. In particular when reading
data stored in parquet files, where it is possible to read only the
data needed, rather than all of all the time and then remove the
one not needed.